+++ packages/twenty-server/src/engine/api/rest/core-query-builder/utils/__tests__/check-fields.utils.spec.ts
@@ -1,0 +1,16 @@
+import { objectMetadataItemMock } from 'src/engine/api/__mocks__/object-metadata-item.mock';
+import { checkFields } from 'src/engine/api/rest/core-query-builder/utils/check-fields.utils';
+
+describe('checkFields', () => {
+  it('should check field types', () => {
+    expect(() =>
+      checkFields(objectMetadataItemMock, ['fieldNumber']),
+    ).not.toThrow();
+
+    expect(() => checkFields(objectMetadataItemMock, ['wrongField'])).toThrow();
+
+    expect(() =>
+      checkFields(objectMetadataItemMock, ['fieldNumber', 'wrongField']),
+    ).toThrow();
+  });
+});
