+++ packages/twenty-server/src/engine/api/rest/core-query-builder/utils/__tests__/get-field-type.utils.spec.ts
@@ -1,0 +1,11 @@
+import { objectMetadataItemMock } from 'src/engine/api/__mocks__/object-metadata-item.mock';
+import { FieldMetadataType } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
+import { getFieldType } from 'src/engine/api/rest/core-query-builder/utils/get-field-type.utils';
+
+describe('getFieldType', () => {
+  it('should get field type', () => {
+    expect(getFieldType(objectMetadataItemMock, 'fieldNumber')).toEqual(
+      FieldMetadataType.NUMBER,
+    );
+  });
+});
